Added suryakamalog_FlyodTrianglePattern.cpp
Floyd's triangle is a right-angled triangular array of natural numbers, used in computer science education. It is named after Robert Floyd. It is defined by filling the rows of the triangle with consecutive numbers, starting with a 1 in the top left corner. 
Given n, print Flyod's Triangle upto n rows.
eg. n = 5
Output will be
1 
2 3 
4 5 6 
7 8 9 10 
11 12 13 14 15